merging fixes into 3d-map

+++ Elements.pde
@@ -79,7 +79,7 @@ class Tooltip extends Element{
     }
     setText(t);
   }
-  void setAttacking(float chance){
+  void setAttacking(int chance){
     JSONObject jo = gameData.getJSONObject("tooltips");
     setText(String.format(jo.getString("attacking"), chance));
   }
